fast_api/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
databases==0.9.0
exceptiongroup==1.2.2
fastapi==0.110.0
greenlet==3.1.1
idna==3.10
loguru==0.7.2
pydantic==1.10.7
pydantic_core==2.16.3
pyhumps==3.8.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.36.3
typing_extensions==4.12.2
